refactor: InventoryBuilder.php
